refactor: 커스텀 구분자 Service 를 일반 구분자 Service 와 분리
- 커스텀 구분자 Service 가 커스텀 구분자와 관계없는 역할도 하고 있음.
- 일반 구분자 Service 와 Custom 구분자 Service 로 분리.
