+++ src/main/java/calculator/Application.java
@@ -6,6 +6,7 @@ import calculator.delimiter.factory.DelimiterFactory;
 import calculator.delimiter.pattern.CustomDelimiterPatternMatcher;
 import calculator.delimiter.pattern.DefaultCustomDelimiterPatternMatcher;
 import calculator.delimiter.service.CustomDelimiterService;
+import calculator.delimiter.service.DelimiterService;
 import calculator.delimiter.validator.CustomDelimiterValidator;
 import calculator.delimiter.validator.DefaultCustomDelimiterValidator;
 import calculator.util.integer.IntegerUtils;
@@ -20,14 +21,17 @@ public class Application {
     private static final CustomDelimiterPatternMatcher customDelimiterPatternMatcher = new DefaultCustomDelimiterPatternMatcher();
     private static final CustomDelimiterValidator customDelimiterValidator = new DefaultCustomDelimiterValidator();
     private static final DelimiterFactory delimiterFactory = new DefaultDelimiterFactory(customDelimiterValidator);
-    private static final CustomDelimiterService customDelimiterService = new CustomDelimiterService(delimiterFactory, customDelimiterPatternMatcher);
+    private static final CustomDelimiterService customDelimiterService = new CustomDelimiterService(
+            delimiterFactory, customDelimiterPatternMatcher
+    );
+    private static final DelimiterService delimiterService = new DelimiterService(delimiterFactory, customDelimiterService);
 
     public static void main(String[] args) {
 
         System.out.println("덧셈할 문자열을 입력해 주세요.");
         String input = IOConsole.readLine();
 
-        List<String> numberStrings = customDelimiterService.extractNumberStrings(input);
+        List<String> numberStrings = delimiterService.extractNumberStrings(input);
 
         try {
             List<Integer> numbers = IntegerUtils.parsePositiveIntegers(numberStrings);
